Add files via upload
This it the notebook with the analysis of the Titanic dataset from Kaggle. The sinking of the Titanic is one of history's most iconic disasters and the challenge is to predict passenger survival based on factors such as gender, age, and class. In this project, after reviewing the approaches taken by other users, I applied data cleaning and preprocessing techniques to organize and prepare the dataset. I extracted actionable insights by analyzing patterns and relationships among features and developed a machine learning model to predict survival with a strong accuracy score. This project demonstrates the ability to handle end-to-end data workflows—from data exploration and visualization to predictive modeling—while delivering insights that translate data into real-world contexts.